backend/requirements-ai.txt
torch==2.9.0  # Latest available PyTorch version
transformers==4.35.2  # Hugging Face transformers
sentence-transformers==2.2.2  # For embeddings
requests==2.31.0  # For web scraping
beautifulsoup4==4.12.2  # For HTML parsing
PyPDF2==3.0.1  # For PDF processing
python-docx==1.1.0  # For Word document processing
openpyxl==3.1.2  # For Excel processing
Pillow==10.1.0  # For image processing
pytesseract==0.3.10  # For OCR
matplotlib==3.8.2  # For plotting
seaborn==0.13.0  # For statistical plots
plotly==5.17.0  # For interactive charts
svgwrite==1.4.3  # For SVG generation
boto3==1.34.0  # For S3 integration
pdfminer.six==20231228  # Better PDF text extraction
fastapi==0.104.1  # For model inference service API
uvicorn[standard]==0.24.0  # ASGI server
pydantic==2.5.0  # Data validation
psutil==5.9.0  # System monitoring
gputil==1.4.0  # GPU monitoring
